refactor: separated tests from `cli/src/filter.rs`
